requirements.txt
Flask_Migrate
alembic
requests
Flask
flask_sqlalchemy
python_magic
